+++ Packages/com.unity.inputsystem/InputSystem/Actions/Interactions/SlowTapInteraction.cs
@@ -6,10 +6,7 @@ namespace UnityEngine.Experimental.Input.Interactions
     {
         public float duration;
         ////REVIEW: this seems stupid; shouldn't a slow tap just be anything that takes longer than TapTime?
-        public float durationOrDefault
-        {
-            get { return duration > 0.0f ? duration : InputSystem.settings.defaultSlowTapTime; }
-        }
+        public float durationOrDefault => duration > 0.0f ? duration : InputSystem.settings.defaultSlowTapTime;
 
         // If this is non-zero, then if the control is held for longer than
         // this time, the slow tap is not performed when the control is finally