+++ Packages/com.unity.inputsystem/InputSystem/Controls/Processors/StickDeadzoneProcessor.cs
@@ -20,15 +20,8 @@ namespace UnityEngine.Experimental.Input.Processors
         public float min;
         public float max;
 
-        public float minOrDefault
-        {
-            get { return min == 0.0f ? InputSystem.settings.defaultDeadzoneMin : min; }
-        }
-
-        public float maxOrDefault
-        {
-            get { return max == 0.0f ? InputSystem.settings.defaultDeadzoneMax : max; }
-        }
+        public float minOrDefault => min == 0.0f ? InputSystem.settings.defaultDeadzoneMin : min;
+        public float maxOrDefault => max == 0.0f ? InputSystem.settings.defaultDeadzoneMax : max;
 
         public Vector2 Process(Vector2 vector, InputControl control)
         {